test: move TabBar initial nav tests into Given_NavigationHotReload as HR tests
