+++ src/Uno.Extensions.RuntimeTests/Uno.Extensions.RuntimeTests/App.xaml.cs
@@ -18,7 +18,6 @@ public partial class App : Application
 		var navigation_UI_Tests = new Given_RouteNotifier();
 		var navigation_ChainedResult_Tests = new Given_ChainedGetDataAsync();
 		var navigation_TabNavigation_Tests = new Given_TabNavigation();
-		var navigation_TabBarInitial_Tests = new Given_TabBarInitialNavigation();
 	}
 
 	/// <summary>
